app/monitor/tls-monitor/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
websocket-client==1.6.4
requests==2.31.0
python-multipart==0.0.6
